Move estimation of variance-covariance matrix to general utilities
